wp-freizeitkalender-eu: Ship local copy of material-design-icons
The material-design-icons was downloaded from a remote server. This means
that it is unclear what the remote server will ship and it allowed the
remote server to track the visitors of the page.

So it is better for privacy and security to ship a local copy. The only
downsite is the requirement to manually update these files regularly in
case the current version is no longer supported or a security problem was
fixed.

Signed-off-by: Sven Eckelmann <sven@narfation.org>