wp-freizeitkalender-eu: Use APCu to cache freizeitkalender.eu response
The freizeitkalender.eu API is rather slow when retrieving our events. And
all our pages are requesting at least two XML files from the remote server.
This made the whole page feel rather slugish because you could see wait
icons in the sidebar.

But most people are not even trying to change the month in the calender
view. So it is possible to just cache the response from the remote server
for a couple of minutes and directly respond the cached value. This
shouldn't create a lot of incoherences because the events are not changed
that often but improves the user experience a lot.

Signed-off-by: Sven Eckelmann <sven@narfation.org>